working on getting the GUI to open when te package is imported. This should be a parallel option to an executable script that runs the  file.
