+++ sededu/__init__.py
@@ -1,1 +1,10 @@
-sededu.py
+# from sededu.root import RootWindow
+# import sys
+# from PyQt5.QtWidgets import *
+# from PyQt5 import QtGui, QtCore
+
+# print("imported SedEdu!")
+# app = QApplication(sys.argv)
+# root = RootWindow()
+# root.show()
+# sys.exit(app.exec_())
